+++ lib/output/writer/amqp_1.go
@@ -11,6 +11,7 @@ import (
 	"github.com/Jeffail/benthos/v3/lib/log"
 	"github.com/Jeffail/benthos/v3/lib/metrics"
 	"github.com/Jeffail/benthos/v3/lib/types"
+	"github.com/Jeffail/benthos/v3/lib/util/amqp/sasl"
 	btls "github.com/Jeffail/benthos/v3/lib/util/tls"
 )
 
@@ -22,6 +23,7 @@ type AMQP1Config struct {
 	TargetAddress string      `json:"target_address" yaml:"target_address"`
 	MaxInFlight   int         `json:"max_in_flight" yaml:"max_in_flight"`
 	TLS           btls.Config `json:"tls" yaml:"tls"`
+	SASL          sasl.Config `json:"sasl" yaml:"sasl"`
 }
 
 // NewAMQP1Config creates a new AMQP1Config with default values.
@@ -31,6 +33,7 @@ func NewAMQP1Config() AMQP1Config {
 		TargetAddress: "",
 		MaxInFlight:   1,
 		TLS:           btls.NewConfig(),
+		SASL:          sasl.NewConfig(),
 	}
 }
 
@@ -93,7 +96,10 @@ func (a *AMQP1) ConnectWithContext(ctx context.Context) error {
 		err     error
 	)
 
-	opts := []amqp.ConnOption{}
+	opts, err := a.conf.SASL.ToOptFns()
+	if err != nil {
+		return err
+	}
 	if a.conf.TLS.Enabled {
 		opts = append(opts, amqp.ConnTLS(true))
 		opts = append(opts, amqp.ConnTLSConfig(a.tlsConf))